# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Student {
  id: number;
  name: string;
  subject: string;
  /**
     * @minimum 0
     * @maximum 10
     */
  score: number;
  createdAt: string;
  updatedAt: string;
}

export interface StudentInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  subject: string;
  /**
     * @minimum 0
     * @maximum 10
     */
  score: number;
}

export interface StudentUpdate {
  /** @minLength 1 */
  name?: string;
  /** @minLength 1 */
  subject?: string;
  /**
     * @minimum 0
     * @maximum 10
     */
  score?: number;
}

export interface SubjectAverage {
  subject: string;
  average: number;
  count: number;
}

export interface StudentScore {
  name: string;
  score: number;
}

export interface RadarPoint {
  subject: string;
  average: number;
}

export type HeatmapRowSubjects = {[key: string]: number};

export interface HeatmapRow {
  name: string;
  subjects: HeatmapRowSubjects;
}

export interface OverviewAnalytics {
  totalStudents: number;
  overallAverage: number;
  topScore: number;
  subjectAverages: SubjectAverage[];
  studentScores: StudentScore[];
  radarData: RadarPoint[];
  heatmap: HeatmapRow[];
}

export interface TeamMemberAverage {
  name: string;
  average: number;
  contributionCount: number;
}

export interface TeamAnalytics {
  groupAverage: number;
  totalContributors: number;
  subjectAverages: SubjectAverage[];
  memberAverages: TeamMemberAverage[];
  heatmap: HeatmapRow[];
}

export interface StudentReportInput {
  name: string;
  subject: string;
  /**
     * @minimum 0
     * @maximum 10
     */
  score: number;
}

export interface TeamReportInput {
  groupAverage: number;
  totalContributors: number;
  subjectAverages: SubjectAverage[];
  memberAverages: TeamMemberAverage[];
}

export interface AiReport {
  headline: string;
  summary: string;
  recommendations: string[];
  generatedAt: string;
}