# lib/api-zod/src/generated/types/aiReport.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AiReport {
  headline: string;
  summary: string;
  recommendations: string[];
  generatedAt: Date;
}