# lib/api-zod/src/generated/types/heatmapRow.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { HeatmapRowSubjects } from './heatmapRowSubjects';

export interface HeatmapRow {
  name: string;
  subjects: HeatmapRowSubjects;
}