# lib/api-zod/src/generated/types/overviewAnalytics.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { HeatmapRow } from './heatmapRow';
import type { RadarPoint } from './radarPoint';
import type { StudentScore } from './studentScore';
import type { SubjectAverage } from './subjectAverage';

export interface OverviewAnalytics {
  totalStudents: number;
  overallAverage: number;
  topScore: number;
  subjectAverages: SubjectAverage[];
  studentScores: StudentScore[];
  radarData: RadarPoint[];
  heatmap: HeatmapRow[];
}