# lib/api-zod/src/generated/types/studentInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface StudentInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  subject: string;
  /**
     * @minimum 0
     * @maximum 10
     */
  score: number;
}