# lib/api-zod/src/generated/types/studentReportInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface StudentReportInput {
  name: string;
  subject: string;
  /**
     * @minimum 0
     * @maximum 10
     */
  score: number;
}