# lib/api-zod/src/generated/types/subjectAverage.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface SubjectAverage {
  subject: string;
  average: number;
  count: number;
}