# lib/api-zod/src/generated/types/teamAnalytics.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { HeatmapRow } from './heatmapRow';
import type { SubjectAverage } from './subjectAverage';
import type { TeamMemberAverage } from './teamMemberAverage';

export interface TeamAnalytics {
  groupAverage: number;
  totalContributors: number;
  subjectAverages: SubjectAverage[];
  memberAverages: TeamMemberAverage[];
  heatmap: HeatmapRow[];
}