# lib/api-zod/src/generated/types/teamMemberAverage.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface TeamMemberAverage {
  name: string;
  average: number;
  contributionCount: number;
}