# lib/api-zod/src/generated/types/teamReportInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { SubjectAverage } from './subjectAverage';
import type { TeamMemberAverage } from './teamMemberAverage';

export interface TeamReportInput {
  groupAverage: number;
  totalContributors: number;
  subjectAverages: SubjectAverage[];
  memberAverages: TeamMemberAverage[];
}